+++ src/providers/labrika.js
@@ -1,0 +1,18 @@
+/**
+ * labrika.js
+ *
+ * REF: https://labrika.com/help/docs/why-does-your-protection-from-bots-prevent-the-labrika-crawler-from-scanning-the-site
+ */
+
+module.exports = {
+  name: 'Labrika',
+  testAddresses: ['178.32.114.61', '162.55.244.68'],
+  ipv4: {
+    addresses: ['178.32.114.61', '162.55.244.68'],
+    ranges: [],
+  },
+  ipv6: {
+    addresses: [],
+    ranges: [],
+  },
+};