+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog for @headwall/trusted-network-providers
 
+## 1.8.0 :: 2025-07-13
+
+* Added Labrika provider
+* Updated assets with scripts/update-assets.sh
+
 ## 1.7.0 :: 2025-06-02
 
 * Added GetTerms provider